chore: remove old rusoto
