+++ Cargo.toml
@@ -65,17 +65,14 @@ metrics-core = "0.5.2"
 metrics-runtime = "0.13.0"
 
 # Aws
-rusoto_core = { version = "0.41.0", optional = true }
-rusoto_core44 = { package = "rusoto_core", version = "0.44.0", optional = true }
+rusoto_core = { version = "0.44.0", optional = true }
 rusoto_s3 = { version = "0.44.0", optional = true }
 rusoto_logs = { version = "0.44.0", optional = true }
 rusoto_cloudwatch = { version = "0.44.0", optional = true }
 rusoto_kinesis = { version = "0.44.0", optional = true }
-rusoto_credential = { version = "0.41.1", optional = true }
-rusoto_credential44 = { package = "rusoto_credential", version = "0.44.0", optional = true }
+rusoto_credential = { version = "0.44.0", optional = true }
 rusoto_firehose = { version = "0.44.0", optional = true }
-rusoto_sts = { version = "0.41.0", optional = true }
-rusoto_sts44 = { package = "rusoto_sts", version = "0.44.0", optional = true }
+rusoto_sts = { version = "0.44.0", optional = true }
 rusoto_signature = { package = "rusoto_signature", version = "0.44.0", optional = true }
 
 # Tower
@@ -355,16 +352,16 @@ sinks = [
   "sinks-vector",
   "sinks-pulsar"
 ]
-sinks-aws_cloudwatch_logs = ["rusoto_core44", "rusoto_credential44", "rusoto_signature", "rusoto_sts44", "rusoto_logs"]
-sinks-aws_cloudwatch_metrics = ["rusoto_core44", "rusoto_credential44", "rusoto_signature", "rusoto_sts44", "rusoto_cloudwatch"]
-sinks-aws_kinesis_firehose = ["rusoto_core44", "rusoto_credential44", "rusoto_signature", "rusoto_sts44", "rusoto_firehose"]
-sinks-aws_kinesis_streams = ["rusoto_core44", "rusoto_credential44", "rusoto_signature", "rusoto_sts44", "rusoto_kinesis"]
-sinks-aws_s3 = ["bytesize", "rusoto_core44", "rusoto_credential44", "rusoto_signature", "rusoto_sts44", "rusoto_s3", "uuid"]
+sinks-aws_cloudwatch_logs = ["rusoto_core", "rusoto_credential", "rusoto_signature", "rusoto_sts", "rusoto_logs"]
+sinks-aws_cloudwatch_metrics = ["rusoto_core", "rusoto_credential", "rusoto_signature", "rusoto_sts", "rusoto_cloudwatch"]
+sinks-aws_kinesis_firehose = ["rusoto_core", "rusoto_credential", "rusoto_signature", "rusoto_sts", "rusoto_firehose"]
+sinks-aws_kinesis_streams = ["rusoto_core", "rusoto_credential", "rusoto_signature", "rusoto_sts", "rusoto_kinesis"]
+sinks-aws_s3 = ["bytesize", "rusoto_core", "rusoto_credential", "rusoto_signature", "rusoto_sts", "rusoto_s3", "uuid"]
 sinks-blackhole = []
 sinks-clickhouse = ["bytesize"]
 sinks-console = []
 sinks-datadog = []
-sinks-elasticsearch = ["base64", "bytesize", "rusoto_core44", "rusoto_credential44", "rusoto_signature", "rusoto_sts44"]
+sinks-elasticsearch = ["base64", "bytesize", "rusoto_core", "rusoto_credential", "rusoto_signature", "rusoto_sts"]
 sinks-file = []
 sinks-gcp = ["base64", "bytesize", "goauth", "smpl_jwt", "uuid"]
 sinks-honeycomb = ["sinks-http"]
