+++ src/lib.rs
@@ -86,7 +86,6 @@ pub mod kafka;
 pub mod list;
 pub mod metrics;
 pub mod region;
-pub mod region2;
 pub mod runtime;
 pub mod serde;
 pub mod shutdown;
