+++ src/sinks/util/mod.rs
@@ -7,8 +7,6 @@ pub mod retries;
 pub mod retries2;
 #[cfg(feature = "rusoto_core")]
 pub mod rusoto;
-#[cfg(feature = "rusoto_core44")]
-pub mod rusoto2;
 pub mod service;
 pub mod service2;
 pub mod sink;
